layout improvements / bug fixes / fixed over flow issues / fixed all chrome developer issues / revisited color schemas on most pages / improved form flow / added better keyboard support for (most) buttons

+++ frontend/src/components/EventCard.jsx
@@ -44,27 +44,23 @@ const EventCard = ({
             className="blur-sm brightness-75"
           />
         </figure>
-
         <div className="card-body">
           <div className="flex justify-between ">
-
-
             <div className="">
               {isLoading && (
                 <span className="loading loading-spinner loading-lg text-info absolute"></span>
               )}
             </div>
 
-            <div className="">
+            <div>
               {eventId && (
-                <div className="ml-auto mr-auto">
                   <QRCode path={eventId} _size={128} />
-                </div>
               )}
             </div>
 
-            <div className="">
+            <div>
               <svg
+                tabIndex={0}
                 xmlns="http://www.w3.org/2000/svg"
                 height="50"
                 width="50"
@@ -72,7 +68,22 @@ const EventCard = ({
                 fill="currentColor"
                 className="fill-white cursor-pointer hover:fill-blue-700"
                 onClick={async () => {
-                  await handleDelete();
+                  let result = window.confirm(
+                    `Are you sure you want to delete this event?\n\nDeleting this event will also delete all responses associated with it.`
+                  );
+                  if (result) {
+                    await handleDelete();
+                  }
+                }}
+                onKeyUp={async (e) => {
+                  if (e.key === "Enter") {
+                    let result = window.confirm(
+                      `Are you sure you want to delete this event?\n\nDeleting this event will also delete all responses associated with it.`
+                    );
+                    if (result) {
+                      handleDelete();
+                    }
+                  }
                 }}
               >
                 <path d="M256 48a208 208 0 1 1 0 416 208 208 0 1 1 0-416zm0 464A256 256 0 1 0 256 0a256 256 0 1 0 0 512zM175 175c-9.4 9.4-9.4 24.6 0 33.9l47 47-47 47c-9.4 9.4-9.4 24.6 0 33.9s24.6 9.4 33.9 0l47-47 47 47c9.4 9.4 24.6 9.4 33.9 0s9.4-24.6 0-33.9l-47-47 47-47c9.4-9.4 9.4-24.6 0-33.9s-24.6-9.4-33.9 0l-47 47-47-47c-9.4-9.4-24.6-9.4-33.9 0z" />
